consolidate component types into one file
