+++ src/platform/packages/shared/shared-ux/document_data_cascade/impl/src/components/data_cascade_impl/data_cascade_row/index.ts
@@ -8,5 +8,4 @@
  */
 
 export { CascadeRowPrimitive } from './cascade_row';
-export type { CascadeRowPrimitiveProps } from './cascade_row';
 export { getCascadeRowNodePath, getCascadeRowNodePathValueRecord } from '../../../lib/utils';
