+++ src/platform/packages/shared/shared-ux/document_data_cascade/impl/src/components/data_cascade_impl/data_cascade_row_cell/index.ts
@@ -8,4 +8,3 @@
  */
 
 export { CascadeRowCellPrimitive } from './cascade_row_cell';
-export type { CascadeRowCellPrimitiveProps } from './cascade_row_cell';
