+++ src/platform/packages/shared/shared-ux/document_data_cascade/impl/src/components/data_cascade_impl/index.ts
@@ -8,8 +8,4 @@
  */
 
 export { DataCascadeImpl, DataCascadeRow, DataCascadeRowCell } from './data_cascade_impl';
-export type {
-  DataCascadeImplProps,
-  DataCascadeRowProps,
-  DataCascadeRowCellProps,
-} from './data_cascade_impl';
+export type { DataCascadeImplProps, DataCascadeRowProps, DataCascadeRowCellProps } from './types';
